Produce forecasts from a vital model

The forecast function allows you to produce future predictions of a functional time series model, where the response is a function of age. If the response variable has been transformed in the model formula, the transformation will be automatically back-transformed and bias-adjusted. More details about transformations in the fable framework can be found in `vignette("transformations", package = "fable")`.

Usage

# S3 method for model_lc
forecast(
  object,
  new_data,
  bootstrap = FALSE,
  times = 5000,
  se = c("innovdrift", "innovonly"),
  jumpchoice = c("fit", "actual"),
  ...
)

# S3 method for model_fmean
forecast(object, new_data, bootstrap = FALSE, times = 5000, ...)

Arguments

object

A mable containing one or more models.

new_data

A `tsibble` containing future information used to forecast.

bootstrap

If `TRUE`, then forecast distributions are computed using simulation with resampled errors.

times

The number of sample paths to use in estimating the forecast distribution when `bootstrap = TRUE`.

se

Method used for computation of standard error. Possibilities: “innovdrift” (innovations and drift) and “innovonly” (innovations only).

jumpchoice

Method used for computation of jumpchoice. Possibilities: “actual” (use actual rates from final year) and “fit” (use fitted rates). The original Lee-Carter method used 'fit' (the default), but Lee and Miller (2001) and most other authors prefer 'actual' (the default).

...

Additional arguments not used. #'

Value

A fable containing the following columns: - `.model`: The name of the model used to obtain the forecast. Taken from the column names of models in the provided mable. - The forecast distribution. The name of this column will be the same as the dependent variable in the model(s). If multiple dependent variables exist, it will be named `.distribution`. - Point forecasts computed from the distribution using the functions in the `point_forecast` argument. - All columns in `new_data`, excluding those whose names conflict with the above.

Details

The forecasts returned contain both point forecasts and their distribution. A specific forecast interval can be extracted from the distribution using the [`hilo()`] function, and multiple intervals can be obtained using [`report()`]. These intervals are stored in a single column using the `hilo` class, to extract the numerical upper and lower bounds you can use [`unpack_hilo()`].