Compute total fertility rate from age-specific fertility rates

Total fertility rate is the expected number of babies per woman in a life-time given the fertility rate at each age of a woman's life.

Usage

total_fertility_rate(.data, fertility)

Arguments

.data

A vital object including an age variable and a variable containing fertility rates.

fertility

Variable in `.data` containing fertility rates. If omitted, the variable with name `fx`, `Fertility` or `Rate` will be used (not case sensitive).

Value

A vital object with total fertility in column `tfr`.

Author

Rob J Hyndman

Examples

# Compute Australian total fertility rates over time
aus_fertility |>
  total_fertility_rate()
#> # A vital: 86 x 4 [1Y]
#>     Year Fertility Exposure  Births
#>    <int>     <dbl>    <dbl>   <dbl>
#>  1  1921      3.11  1418909 136874.
#>  2  1922      3.10  1450407 138845.
#>  3  1923      3.01  1482482 137207.
#>  4  1924      2.97  1514008 137369.
#>  5  1925      2.95  1547341 138818.
#>  6  1926      2.85  1579812 136442.
#>  7  1927      2.81  1613370 136994.
#>  8  1928      2.77  1645999 137266.
#>  9  1929      2.64  1673581 132331.
#> 10  1930      2.58  1696345 130890.
#> # ℹ 76 more rows