Compute period life tables from age-specific mortality rates

All available years and ages are included in the tables. $qx = mx/(1 + ((1-ax) * mx))$ as per Chiang (1984). Warning: the code has only been tested for data based on single-year age groups.

Usage

life_table(.data, mortality)

Arguments

.data

A vital including an age variable and a variable containing mortality rates.

mortality

Variable in .data containing Mortality rates (mx). If omitted, the variable with name mx, Mortality or Rate will be used (not case sensitive).

Value

A vital object containing the index, keys, and the new life table variables mx, qx, lx, dx, Lx, Tx and ex.

References

Chiang CL. (1984) The life table and its applications. Robert E Krieger Publishing Company: Malabar.

Keyfitz, N, and Caswell, H. (2005) Applied mathematical demography, Springer-Verlag: New York.

Preston, S.H., Heuveline, P., and Guillot, M. (2001) Demography: measuring and modeling population processes. Blackwell

Author

Rob J Hyndman

Examples

# Compute Victorian life table for females in 2003
aus_mortality |>
  dplyr::filter(Code == "VIC", Sex == "female", Year == 2003) |>
  life_table()
#> # A vital: 101 x 15 [?]
#> # Key:     Age x (Sex, State, Code) [101 x 1]
#>     Year   Age Sex   State Code       mx      qx    lx      dx    Lx    Tx    ex
#>    <int> <int> <chr> <chr> <chr>   <dbl>   <dbl> <dbl>   <dbl> <dbl> <dbl> <dbl>
#>  1  2003     0 fema… Vict… VIC   4.12e-3 4.10e-3 1     4.10e-3 0.996  83.2  83.2
#>  2  2003     1 fema… Vict… VIC   5.76e-4 5.76e-4 0.996 5.73e-4 0.996  82.2  82.6
#>  3  2003     2 fema… Vict… VIC   2.03e-4 2.03e-4 0.995 2.02e-4 0.995  81.2  81.6
#>  4  2003     3 fema… Vict… VIC   6.65e-5 6.65e-5 0.995 6.61e-5 0.995  80.2  80.6
#>  5  2003     4 fema… Vict… VIC   1.32e-4 1.32e-4 0.995 1.31e-4 0.995  79.2  79.6
#>  6  2003     5 fema… Vict… VIC   3.31e-5 3.31e-5 0.995 3.30e-5 0.995  78.3  78.6
#>  7  2003     6 fema… Vict… VIC   0       0       0.995 0       0.995  77.3  77.7
#>  8  2003     7 fema… Vict… VIC   6.46e-5 6.46e-5 0.995 6.42e-5 0.995  76.3  76.7
#>  9  2003     8 fema… Vict… VIC   3.17e-5 3.17e-5 0.995 3.15e-5 0.995  75.3  75.7
#> 10  2003     9 fema… Vict… VIC   3.12e-5 3.12e-5 0.995 3.11e-5 0.995  74.3  74.7
#> # ℹ 91 more rows
#> # ℹ 3 more variables: rx <dbl>, nx <dbl>, ax <dbl>