Functions to smooth demographic data

These smoothing functions allow smoothing of a variable in a vital object. The vital object is returned along with some additional columns containing information about the smoothed variable: usually .smooth containing the smoothed values, and .smooth_se containing the corresponding standard errors.

Usage

smooth_spline(.data, .var, k = -1, weights = NULL)

smooth_mortality(.data, .var, b = 65, power = 0.4, k = 30, weights = NULL)

smooth_fertility(.data, .var, lambda = 1e-10)

smooth_loess(.data, .var, span = 0.2, weights = NULL)

Arguments

.data

A vital object

.var

name of variable to smooth

k

Number of knots to use for penalized regression spline estimate.

weights

Vector of weights.

b

Lower age for monotonicity. Above this, the smooth curve is assumed to be monotonically increasing.

power

Power transformation for age variable before smoothing. Default is 0.4 for mortality data and 1 (no transformation) for fertility or migration data.

lambda

Penalty for constrained regression spline.

span

Span for loess smooth.

Value

vital with added columns containing smoothed values and their standard errors

Details

smooth_mortality() use penalized regression splines applied to log mortality with a monotonicity constraint above age b. The methodology is based on Wood (1994). smooth_fertility() uses weighted regression B-splines with a concavity constraint, based on He and Ng (1999). The function smooth_loess() uses locally quadratic regression, while smooth_spline() uses penalized regression splines.

References

Hyndman, R.J., and Ullah, S. (2007) Robust forecasting of mortality and fertility rates: a functional data approach. Computational Statistics & Data Analysis, 51, 4942-4956. https://robjhyndman.com/publications/funcfor/

Author

Rob J Hyndman

Examples

library(dplyr)
#> 
#> Attaching package: ‘dplyr’
#> The following objects are masked from ‘package:stats’:
#> 
#>     filter, lag
#> The following objects are masked from ‘package:base’:
#> 
#>     intersect, setdiff, setequal, union
aus_mortality |>
  filter(State == "Victoria", Sex == "female") |>
  smooth_mortality(Mortality)
#> # A vital: 12,120 x 10 [1Y]
#> # Key:     Age, Sex, State, Code [101]
#>     Year   Age Sex    State   Code  Mortality Exposure Deaths .smooth .smooth_se
#>    <int> <int> <chr>  <chr>   <chr>     <dbl>    <dbl>  <dbl> <dbl[1>  <dbl[1d]>
#>  1  1901     0 female Victor… VIC     0.100      13993   1405 0.102     0.0147  
#>  2  1901     1 female Victor… VIC     0.0235     13079    308 0.0233    0.00327 
#>  3  1901     2 female Victor… VIC     0.00806    12409    100 0.00827   0.00106 
#>  4  1901     3 female Victor… VIC     0.00472    12931     61 0.00479   0.000559
#>  5  1901     4 female Victor… VIC     0.00370    12986     48 0.00370   0.000397
#>  6  1901     5 female Victor… VIC     0.00324    13589     44 0.00330   0.000330
#>  7  1901     6 female Victor… VIC     0.00310    13872     43 0.00308   0.000300
#>  8  1901     7 female Victor… VIC     0.00284    14077     40 0.00285   0.000266
#>  9  1901     8 female Victor… VIC     0.00261    14198     37 0.00256   0.000233
#> 10  1901     9 female Victor… VIC     0.00225    14694     33 0.00226   0.000199
#> # ℹ 12,110 more rows
aus_fertility |>
  smooth_fertility(Fertility)
#> # A vital: 3,010 x 7 [1Y]
#> # Key:     Age [35]
#>     Year   Age Fertility Exposure Births .smooth .smooth_se
#>    <int> <int>     <dbl>    <dbl>  <dbl>   <dbl>      <dbl>
#>  1  1921    15   0.00175    49228   86.1 0.00114    0.0149 
#>  2  1921    16   0.00685    47779  327.  0.00685    0.00996
#>  3  1921    17   0.0183     45894  839.  0.0193     0.0100 
#>  4  1921    18   0.0399     45683 1824.  0.0399     0.00944
#>  5  1921    19   0.0676     43364 2931.  0.0665     0.00982
#>  6  1921    20   0.0914     47016 4297.  0.0929     0.00913
#>  7  1921    21   0.120      48023 5787.  0.118      0.00978
#>  8  1921    22   0.142      46368 6566.  0.143      0.00916
#>  9  1921    23   0.16       45123 7220.  0.160      0.00917
#> 10  1921    24   0.167      46709 7782.  0.167      0.00978
#> # ℹ 3,000 more rows